+++ frontend/src/app/programs/observations/form/form.component.ts
@@ -3,7 +3,7 @@ import { geometryValidator, ngbDateMaxIsToday } from './formValidators';
 import { MainConfig } from './../../../../conf/main.config';
 import { BaseLayer } from '../../programs.models';
 import * as L from 'leaflet';
-
+import { getPreferredName } from '../../../api/getPreferredName';
 import {
     AfterViewInit,
     Component,
@@ -20,6 +20,7 @@ import {
     distinctUntilChanged,
     map,
     share,
+    switchMap,
     tap,
 } from 'rxjs/operators';
 import { FeatureCollection } from 'geojson';
@@ -42,6 +43,7 @@ import { MapService } from '../../base/map/map.service';
 import { GNCFrameworkComponent } from '../../base/jsonform/framework/framework.component';
 import { RefGeoService } from '../../../api/refgeo.service';
 import { ControlPosition } from 'leaflet';
+import { TaxhubService } from '../../../api/taxhub.service';
 
 // declare let $: any;
 
@@ -79,7 +81,9 @@ const map_conf = {
 const taxonSelectInputThreshold = MainConfig.taxonSelectInputThreshold;
 const taxonAutocompleteInputThreshold =
     MainConfig.taxonAutocompleteInputThreshold;
-const taxonAutocompleteFields = MainConfig.taxonAutocompleteFields;
+
+//  TODO: [LIMIT100-TAXON-COMPAT-THV2] normalement à enlever
+// const taxonAutocompleteFields = MainConfig.taxonAutocompleteFields;
 const taxonAutocompleteMaxResults = 10;
 const taxonDisplayImageWhenUnique =
     'taxonDisplayImageWhenUnique' in MainConfig
@@ -148,6 +152,8 @@ export class ObsFormComponent implements AfterViewInit {
     photos: any[] = [];
     existing_photos: any[] = [];
     mapVars: any = {};
+    isInvalidDirty = false;
+    loading: boolean = true;
 
     constructor(
         @Inject(LOCALE_ID) readonly localeId: string,
@@ -158,10 +164,12 @@ export class ObsFormComponent implements AfterViewInit {
         private toastr: ToastrService,
         private auth: AuthService,
         private mapService: MapService,
-        private _refGeoService: RefGeoService
+        private _refGeoService: RefGeoService,
+        private _taxhubService: TaxhubService
     ) {}
 
     ngOnInit(): void {
+        this._taxhubService.loadAndCacheData();
         this.program_id = this.data.program_id;
         this.coords = this.data.coords;
         this.updateMunicipality();
@@ -185,41 +193,51 @@ export class ObsFormComponent implements AfterViewInit {
     }
 
     ngAfterViewInit() {
+        this.loading = true;
         this.programService
             .getProgram(this.program_id)
             .subscribe((result: FeatureCollection) => {
                 this.program = result;
                 this.taxonomyListID =
                     this.program.features[0].properties.taxonomy_list;
-                this.surveySpecies$ = this.programService
-                    .getProgramTaxonomyList(this.taxonomyListID)
+                    this.surveySpecies$ = this.programService
+                    .getAllProgramTaxonomyList()
                     .pipe(
-                        tap((species) => {
-                            this.taxa = species;
-                            this.taxaCount = Object.keys(this.taxa).length;
-                            if (
-                                this.taxaCount >=
-                                this.taxonAutocompleteInputThreshold
-                            ) {
-                                this.inputAutoCompleteSetup();
-                            } else if (this.taxaCount == 1) {
-                                this.onTaxonSelected(this.taxa[0]);
-                            }
-                        }),
-                        share()
+                      map((listsTaxonomy) => {
+                        this.taxaCount = listsTaxonomy
+                          .filter((lt) => lt.id_liste === this.taxonomyListID)
+                          .map((lt) => lt.nb_taxons)[0];
+                        return this.taxaCount <this.taxonAutocompleteInputThreshold;
+                      }),
+                      switchMap((shouldFetchTaxonomyList) => {
+                        if (shouldFetchTaxonomyList) {
+                          return this.programService.getProgramTaxonomyList(this.taxonomyListID,{"limit": this.taxonAutocompleteInputThreshold}).pipe(
+                            tap((species) => {
+                              this.taxa = this._taxhubService.setMediasAndAttributs(species);
+                            }),
+                            map((species: TaxonomyList) => {
+                              if (this.taxaCount < this.taxonAutocompleteInputThreshold) {
+                                return species.sort((a, b) => {
+                                  const taxA = a.nom_francais || a.taxref.nom_vern || '';
+                                  const taxB = b.nom_francais || b.taxref.nom_vern || '';
+                                  return taxA.localeCompare(taxB);
+                                });
+                              } else {
+                                return species;
+                              }
+                            })
+                          );
+                        } else {
+                          this.loading = false;
+                          return [];
+                        }
+                      }),
+                      share()
                     );
-                this.surveySpecies$.subscribe((res: TaxonomyList) => {
-                    res.sort((a, b): number => {
-                        const tax_a = a.nom_francais
-                            ? a.nom_francais
-                            : a.taxref.nom_vern;
-                        const tax_b = b.nom_francais
-                            ? b.nom_francais
-                            : b.taxref.nom_vern;
-                        return tax_a.localeCompare(tax_b);
-                    });
-                    this.surveySpecies = res;
-                });
+                        this.surveySpecies$.subscribe((sortedSpecies) => {
+                            this.surveySpecies = sortedSpecies;
+                            this.loading = false;
+                          });
 
                 if (this.program.features[0].properties.id_form) {
                     // Load custom form if one is attached to program
@@ -431,7 +449,6 @@ export class ObsFormComponent implements AfterViewInit {
     }
 
     patchForm(updateData) {
-        // console.log("updateData", updateData)
         const taxon = updateData.taxon || {
             media: updateData.taxref.media_url,
             taxref: updateData.taxref,
@@ -449,51 +466,9 @@ export class ObsFormComponent implements AfterViewInit {
         });
     }
 
-    inputAutoCompleteSetup = () => {
-        for (let taxon in this.taxa) {
-            for (let field of taxonAutocompleteFields) {
-                if (this.taxa[taxon]['taxref'][field]) {
-                    this.species.push({
-                        name:
-                            field === 'cd_nom'
-                                ? `${this.taxa[taxon]['taxref']['cd_nom']} - ${this.taxa[taxon]['taxref']['nom_complet']}`
-                                : this.taxa[taxon]['taxref'][field],
-                        cd_nom: this.taxa[taxon]['taxref']['cd_nom'],
-                        icon:
-                            this.taxa[taxon]['medias'].length >= 1
-                                ? // ? this.taxa[taxon]["medias"][0]["url"]
-                                  MainConfig.API_TAXHUB +
-                                  '/tmedias/thumbnail/' +
-                                  this.taxa[taxon]['medias'][0]['id_media'] +
-                                  '?h=20'
-                                : 'assets/default_image.png',
-                    });
-                }
-            }
-        }
-        this.autocomplete = 'isOn';
-    };
+    // Expose to HTML
+    getPreferredName = getPreferredName;
 
-    inputAutoCompleteSearch = (text$: Observable<string>) =>
-        text$.pipe(
-            debounceTime(200),
-            distinctUntilChanged(),
-            map((term) =>
-                term === '' // term.length < n
-                    ? []
-                    : this.species
-                          .filter(
-                              (v) =>
-                                  v['name']
-                                      .toLowerCase()
-                                      .indexOf(term.toLowerCase()) > -1
-                              // v => new RegExp(term, "gi").test(v["name"])
-                          )
-                          .slice(0, taxonAutocompleteMaxResults)
-            )
-        );
-
-    inputAutoCompleteFormatter = (x: { name: string }) => x.name;
     disabledDates = (date: NgbDate, current: { month: number }) => {
         const date_impl = new Date(date.year, date.month - 1, date.day);
         return date_impl > this.today;
@@ -503,7 +478,7 @@ export class ObsFormComponent implements AfterViewInit {
         this.selectedTaxon = taxon;
         this.obsForm.controls['cd_nom'].patchValue({
             cd_nom: taxon.taxref['cd_nom'],
-            name: taxon.taxref.nom_complet,
+            name: getPreferredName(taxon),
         });
     }
 
@@ -541,14 +516,10 @@ export class ObsFormComponent implements AfterViewInit {
         if (isNaN(cd_nom)) {
             cd_nom = Number.parseInt(taxon.cd_nom);
         }
-        // Need to convert the defined interface to an array to have
-        // access to the filter function
-        const tempTaxa = this.taxa as Array<unknown> as Array<TempTaxa>;
-        const taxon_name: TempTaxa = tempTaxa.filter(
-            (t) => t.cd_nom == cd_nom
-        )[0];
+        // const taxon_name = this.selectedTaxon.nom_francais ? this.selectedTaxon.nom_francais : this.selectedTaxon.taxref.nom_vern;
+        const taxon_name = taxon.name
         formData.append('cd_nom', cd_nom.toString());
-        formData.append('name', taxon_name.nom_francais);
+        formData.append('name', taxon_name);
         const obsDateControlValue = NgbDate.from(
             this.obsForm.controls.date.value
         );
@@ -662,5 +633,27 @@ export class ObsFormComponent implements AfterViewInit {
         let resp = this.obsForm.valid;
         if (this.customForm.json_schema) resp = resp && this.jsonValid;
         return resp;
+    }
+
+
+    onSelectedTaxon(taxon) {
+        this.programService
+            .getTaxonInfoByCdNom(taxon.item['cd_nom'])
+            .subscribe((taxonFullInfo) => {
+                const taxonWithTaxhubInfos= this._taxhubService.setMediasAndAttributs(taxonFullInfo)
+                this.selectedTaxon = taxonWithTaxhubInfos[0];
+                this.obsForm.controls['cd_nom'].patchValue({
+                    cd_nom: this.selectedTaxon['cd_nom'],
+                    name: getPreferredName(this.selectedTaxon),
+                    icon:
+                        this.selectedTaxon['medias'].length >= 1
+                            ? // ? this.taxa[taxon]["medias"][0]["url"]
+                              MainConfig.API_TAXHUB +
+                              '/tmedias/thumbnail/' +
+                              this.selectedTaxon['medias']['id_media'] +
+                              '?h=20'
+                            : 'assets/default_image.png',
+                });
+            });
     }
 }